wi9p wi9m: Corrected antenna map constants for wi9m
This is a fix for Vantive 33068.  The antenna maps for the wi9m
were incorrectly updated when we fixed the maps for the wi9p.
This update changes the code so that different antenna maps are
used for the different platforms and corrects the map for the
wi9m
x
Signed-off-by: Charles Gordon <charles.gordon@digi.com>