wi9p wifi: Fixed bug where retry info was zeroed when frames enqueued
Found that the retry information for frames was being zeroed out when
the frames were enqueued, instead of when they were dequeued.  This
could cause the retry information for the currently transmitting frame
to be zeroed out if another frame was enqueued while it was in progress.
Fixed it so that the retry info is cleared when the frame is dequeued
for processing, not when it is enqueued.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>