fim-sdio: allow to use two FIMs for SDIO
SDIO functionality could only be used on one FIM because it
was assumed that the user was using the FIM Application Kit
hardware board and in this board the Card Detect (CD) and
Write Protect (WP) signals were shared between both FIMs, apart
from the fact that there is only one SD socket.

Customers however might use a custom hardware and might want to
use or not the CD and WP functionality, or they may use different
GPIOs for them. The following patch removes the old dependencies
by adding new configuration items and letting the user decide
whether he will use the FIM Application Kit hardware board or not
and, if not, let him choose whether he wants CD/WP functionality
for each FIM and select the GPIOs and GPIOs funcs he will be using
for them.

If built as a kernel module, this can be specified with six new
arguments:
	'cd0'      -> for Card Detect on FIM0
	'cd0_func' -> for Card Detect GPIO function on FIM0
	'wp0'      -> for Write Protect on FIM0
	'cd1'      -> for Card Detect on FIM1
	'cd1_func' -> for Card Detect GPIO function on FIM1
	'wp1'      -> for Write Protect on FIM1

If a parameter is not specified, the corresponding functionality
won't be available for that FIM.

If Card Detect functionality is not available, it is assumed the card is
always plugged.
If Write Protect functionality is not available, it is
assumed Write Protect is off.

This patch also fixes a bug when the CD was used as normal GPIO. In this
case, a kernel timer was being used for polling the CD gpio, but when the
module was removed, the timer might still be called. A boolean variable
was created to protect the timer timeout function from doing anything if
the timer was disabled.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>