spi: spidev_test gives error upon 1-byte transfer
The sample application spidev_test.c is using SPI_IOC_MESSAGE
ioctl to do an SPI transfer. This ioctl returns the number of
bytes successfully transmitted or a negative error code upon
erroneous completion. The application however is returning an
error if the result of the ioclt if the return value is 1.
This makes the application to fail upon 1-byte length transfers.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>