serial: add the serial ports names to kernel UART report
This was a change request by Thomas Kuhlmann (#33634) because
not always the tty index matches the port number in the CPU (for
example in NS9360, ttyS0 is PORT B and ttyS1 is PORT A).

The patch adds the label with the port name to the line where the
kernel reports the detected UARTs.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>