mxc_ts: improved wait on irq support, avoding a race condition
This commit improves (but not fixes) a problem detected with the touch
driver, that was missing an interrupt condition (a pen down after a pen up)
and was causing that the driver were not able to detect additional interrupts
if the pen was down, because there were no new interrupts generated.

Thsi workarounds the problem by forcing the driver to wait on interrupts only
after 2 consecutive pen-up samples. This seems to fix the problem, but its
not 100 sure because the problem has not been analyzed with enough detail.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>