mxc_nd2: disable writing BBT into flash
We need to revisit in future the way used to handle bad blocks,
using Bad Block Tables or just the informatino on the spare area.
With the current linux driver, when writing the BBT into flash, causes
that the bootlaoder missunderstands the 2 sectors where the BBT is stored,
considering them as bad blocks. So, at the moment we skip the writing of
the BBT into the flash.

This review needs to be done in parallel with the bootloader (and WinCE).

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>