pmic_adc: implement shutdown to disable the adc and solve reboot problem
This commit implements the shutdown function to solve a problem that was
happening when the touchscreen was enabled (and had been open).
In that scenario, the reboot shell command was not working (the system
was not rebooting properly, stalling after Restarting system message).

Stopping the ADC by calling the deinit function seems to fix the problem.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>