ccwmx51: Do not unconditionally set the PMIC as a suspend wake up source
This behaviour was causing that even when the user has not explicitely set
a wakeup source the system was suspending without a way to awake. With
this change the RTC wake up source is set up when its suspend method is
called, making it possible to check for enabled wake up sources on the
platform suspend enter.

Signed-off-by: Alejandro Gonzalez <alex.gonzalez@digi.com>