ccwmx51: reset external module peripherals
This commit resets all the (built in the kernel) external module
peripherals (external mac, fec phy and wireless module).
Recent versions of u-boot doesnt touch the reset lines if the
peripherals are not used by the boot loader and the kernel needs
to set to high the peripherals reset lines.
This code should be move to a different place (peripheral
initialization code would be preferred)

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>