cfi: disable write-buffer on Numonyx flash M29W640
For some strange reason, this Numonyx flash should support write-buffer
operations but the CFI driver complains with a TIMEOUT error message
on every write attempt.
This patch disables the write-buffer algorithm for M29W640 flashes
as a temporary workaround.

Note that this patch has a performance impact of around 50% on write
accesses to flash.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>