ENGR00125259 - EPDC fb: Random display artifacts in unit tests
Random artifacts were caused by corruption of the EPDC working
buffer.  This occurred because the working buffer was being
allocated too small.  This meant other accesses to FB-maintained
buffers was corrupting the working buffer and causing random
data to be drawn to the display.  Fixing the working buffer size
causes the artifacts to disappear.

Signed-off-by: Danny Nold <dannynold@freescale.com>