wi9p wifi: Change to spike suppression routine requested by Mike Schaffner
The original routine determined when the timing measurement was finished
by anding the MAC_STATUS register with 0xff00 and comparing it to 0x100.
According to Mike, the upper nibble will never be zero and this will cause
the routine to loop until it finally times out.  The fix is to and the
register with 0x100 instead.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>