ENGR00126331 MX28 ALSA: fix 11Khz and 22KHz stream playback issue
For playing 11Khz and 22Khz stream, currently the DMA may be in
an abnormal state during mxs_pcm_close phase which causes the
SAIF clock to be disabed unexpectedly.
Thereafter audio codec will be unable work again and system hangs.

Add reset operation for DMA before release it to make sure in a
right state to avoid this issue.

Signed-off-by: Aisheng.Dong <b29396@freescale.com>