ENGR00127147: MX50: Add support for Automatic Low Power Mode on DDR
Configure the Databahn so that the DDR automatically enters Low Power Mode 4.
This will allow DDR to enter into self-refresh after an idle timeout.
Also added code to disable DDR_CLKGAT bits in the CCM when no bus masters
are accessing the DDR and ARM is in WFI.
Removed calls to change DDR frequency to 24MHz in LPAPM mode as this is
causing some random lockups. DDR is left at 266MHz in LPAPM mode currently.

Signed-off-by: Ranjani Vaidyanathan <ra5478@freescale.com>