Configure CSPI_RDY as GPIO
On the JSK boards, if we leave this pin as default, it will be configured
by the SPI driver and the target will halt on PMIC initial access. Setting
this pin as GPIO with a pull-up avoids the problem.

This pin is used as CS for the LCD1 on the EAK boards and as CS for LCD1
and LCD2 on the JSK boards.

Signed-off-by: Alejandro Gonzalez <alex.gonzalez@digi.com>