ENGR00131968 esdhc: fix esdhc gets CRC errors when AHB is 24Mhz
This issue is found on MX50 ARM2 board when AHB is 24Mhz and esdhc
is 50Mhz (maybe just need > 24Mhz, however not tested).

We found it's related to the clock auto gate function of esdhc which
enables the controller to automatically gate on/off the required clocks
based on its internal logic judgement.

After disabling clock auto gate function, we could not reproduce it anymore.
Thus we disable clock auto gate here to get better compatibility.

It could be regarded as a workaround due to it may be h/w bug of auto gate
of esdhc(not confirmed).

Signed-off-by: Aisheng.Dong <b29396@freescale.com>