ccwmx51: add support to swap the bbi in the nand
This commit adds the support to swap the bad block indicator.
This is needed due to the special way that the nfc works,
not being able of writing the full data page and then the
spare area, on 2K page devices.
Because the nfc writes blocks of 512+16 bytes, the bad block
has to be relocated to take advantage of the hardware support
when writing/reading to/from the nand using the ECC.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>