Fix crash on v4l2 application close
The POSIX pid is the tgid (thread group id), not the pid, which is just
the internal scheduler thead id. mxc_streamoff should always be called
for the running tgid, otherwise the csi callback is not disabled and
sometimes we have a kernel dump.

Probably storing the tgid instead of the pid is more correct, but on
process startup they are equal so it's not strictly needed.

Signed-off-by: Alejandro Gonzalez <alex.gonzalez@digi.com>