ENGR00133593 ePXP: make clk_disable() be called only in process context
The main reason is to allow for voltage scaling which will
happen as a result of calls to clk_[en|dis]able().
Voltage scaling is done using the Regulator API, which talks to PMIC
using SPI. SPI code can sleep and hence clk_[en|dis]able() cannot be
called in an ISR/tasklet etc contexts.

Signed-off-by: Robby Cai <R63905@freescale.com>