pwm: autoadjust prescaler for period calculation
Calculate prescaler when the period is set. The formula is only valid
if other timers sharing the prescaler are not running.
When the period increases, the prescaler increases as well. The original
duty cycle may be small and will increase when tick time increases with
the prescaler. If the period is reduced later the original duty is not
restored.
E.g.
- Period is 20ms and duty is 1ms.
- Period increases to 180ms and minimum duty becomes 2ms.
- Period reduces back to 20ms but duty remains 2ms.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>