ENGR38305845-1 NAND : make the mil_ecc_read_oob() reads the whole OOB by default
The origin code reads the whole OOB only when the MTD_OOB_RAW is set.
In the MTD_OOB_PLACE situation, the code will do nothing which will
cause the following command lost the OOB data:
	#nanddump -f f.txt -p -s 0x200000 -l 0x1000 -n /dev/mtd0

So make the function reads the whole OOB by default will fix the
problem.

Signed-off-by: Huang Shijie <b32955@freescale.com>