ccmwx51: workaround random crash at boot time
This commit modifies the code that enables the cpu clock.
Instead of enabling that clock early in the clock configuration
sequence, it is enabled late.
The early enable was causing a random crash at boot time on the
ConnectCore Wi-i.MX51 for unknown reasons.
Enabling the cpu clock late, seems to work around the problem.
This issue have to be investigaged to better understand the
real cause of the problem and the real solution.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>