serial: swap device and driver names in samsung serial
This commit will make the serial devices appear as /dev/ttySACx
as specified in 'Documentation/arm/Samsung-S3C24XX/Overview.txt'
instead of populating them as s3c2410_serialx (which is the name
of the driver).
This change must be followed up by a rule change in the file system
to correctly map ttySACx to ttySx.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>