ccwmx51 ipuv3: fix incorrect video on LCD when dual head (#36848)
This commit fixes some problems that where happening on LCD
displays, when dual head was enabled and when tty0 console
was being used (see #36848 for further info).
This was a side effect of some modifications that have been
done on the ipuv3 video driver to improve the visualization
of VGA video/modes. There was a side effect (the problem
described here).
It has been fixed by applying the new setting on the ipuv3
driver exclusivelly when the VGA interface is being used.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>