ccwmx51: modify the way the pmic is accessed (wireless issue)
This commit reverts, partially, a previous commit that was done
to allow the use of the second touch (changing the way we access
the pmic, through the spi interface).
This was causing a strange side effect... the wireless interface
was, randomly, not accessible because it was not initialized
correctly (not loaded or the mac address was not correctly set).

Reverting this changes seems to fix the problem, causing a side
effect on the second touch interface. This problem needs to be
investigated and properly resolved.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>