piper wifi: Fix for Vantive 35641 -- range check txpower level
The original mac80211 code in ieee_hw_config:main.c rejected the power
level 0 dBm (which is actually not really 0), and would accept
negative values.  I changed the code to accept 0 and to change
negative power values to be 0.

Note that it appears the negative power values were passed to the
function due to a bug in iwconfig.  When the user specifies the
tx power level in units of mW, iwconfig passes negative values to
the mac80211 ioctl for some reason.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>
(cherry picked from commit c2c552f96571983769123af7cc9f5d3678ff96e1)

Signed-off-by: Charles Gordon <charles.gordon@digi.com>