ccwmx51: fix crash on 2nd mac if not enabled in uboot
This commit fixes a problem that was happening when the 2nd
eth mac was used and it was not enabled in uboot.
The problem is that the registers to access the external bus
was not properly configured and therefore and exception was
triggered when the system was accessing the bus.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>