ccwmx51: fix different problems related with 2nd touch
This commit fixes different problems related with the 2nd touch.
-Bad performance when using the 2nd touch.
-Reset problem when dual touch was used.

The problem was caused due to the need of using the the spi poll
transfers to access the pmic and the improper implementation of that
function.
The issue has been workarounded by using poll spi transfers on
all the transfers over the spi bus num 1.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>