FIM CAN: Fixed problem with transmitting frames with extended IDs
SA found a problem that had been in the driver since the original version.
We were never able to transmit frames with extended IDs correctly.  The
code was not formatting the control and ID fields of the frame header
correctly when extended IDs were used.  I fixed this.

This was Vantive 37154.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>