fim_serial: deactivate PM functions if built as module
The PM resume function needs to request the firmware from the file
system if the FIM driver is built as a loadable module. The problem
is the file system is not ready during the resuming of the system
leading to problems reenabling the FIM driver.
PM functions work fine when the driver is built-in, so we deactivate
PM functions if the driver is loaded as module.

This fixes vantive #35491.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
(cherry picked from commit e59a3d05c1273529b22336fe290d7b61d17a31ed)

Signed-off-by: Hector Palacios <hector.palacios@digi.com>