ENGR00137489 mx50 rdp: specify the regulator in esdhc2 platform data
By defining the regulator name in the esdhc platform data,
the esdhc driver will dynamically check and set a proper voltage
for the card.
This can fix the issue that SD32 can not work well on mx50 rdp board
because the default voltage of VSD seems can not meet its requirement.

This is only a workaround.
The root cause may be related to HW, board degsin or WiFi card.

Signed-off-by: Aisheng.Dong <b29396@freescale.com>