Piper Wifi: Reduced likelihood of spike suppression failure at startup
Due to a problem in the piper chip, we have to execute a special routine
that is suppose to synchronize the internal timing of different components
in the chip.  This routine usually fails and we have to keep retrying it
until it succeeds.  There is no way to force it to succeed.  Before this
change the driver would retry the spike suppression routine up to 300
times.  With this number of retries we occasionally would see the driver
fail to load when all the retries failed.  I changed the maximum number
of retries to 30000.  Hopefully this will make the failure extremely
rare.  I don't want to have the routine retry indefinitely because that
could cause an infinite loop if there is some sort of hardware problem
(other than the one we are trying to work around).

Signed-off-by: Charles Gordon <charles.gordon@digi.com>