fim can: Fixed problem with restarting the CAN driver
The driver was having problems restarting when it received the commands

    ifconfig can0 down
    ifconfig can0 up

The problem was that the original code did not completely shutdown and
then restart the FIM.  It tried to just pause the FIM and restart.  I
changed the code so that the "ifconfig can0 down" will completely shut
down the fim, and then "ifconfig can0 up" will restart it from scratch.

This should fix Vantives 37251 and 36779.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>
(cherry picked from commit 9f0defe6b17117934457251411d121e44db9e6ea)

Signed-off-by: Hector Palacios <hector.palacios@digi.com>