[ccwmx51]: Correctly configure the base address for the low level UART output
This patch configures the UART base address based on the UART
configuration. The default is UART2, the default console in the CCWMX51JS,
but if that is not enabled it will use UART1 and then UART3. No further
checks regarding whether the console is active or not are done.

On older kernels, the status of the first UART was checked, and if it
was not functional then the next UART was tried. That code is no more
in the BSP.

Signed-off-by: Alejandro Gonzalez <alex.gonzalez@digi.com>