[ccwmx51]: Fix the support for DEBUG_LL
The DEBUG_LL configuration option allows to use assembly functions
(printascii, printhex) to have serial console output before the console
is initialized. This assumes the serial has already been configured by
the bootloader. The BSP did not compile with this option, and once it
did the code disabled the UART clock on clock initialization, missing
any oops which appeared between then and the console being enabled.

Signed-off-by: Alejandro Gonzalez <alex.gonzalez@digi.com>