[ccwmx51]: USB wake up is not currently supported
Freescales's 11.01 BSP includes USB wake up support.

By default it registers USB hosts as wake up sources. This means that
the user no longer controls what are the allowed ways to wake up.

Also, it doesn't allow to cancel the suspend of a device if no wakeup
sources are registered. The system blocks on a flag that should be cleared
by a kernel thread which never runs as the tasks hasn't been thawned yet.

In all it will need more work to make it usable so it is conditionally
compiled out.

Signed-off-by: Alejandro Gonzalez <alex.gonzalez@digi.com>