Wifi Piper: Reverted file to official version
We had patched this file in an earlier release for Vantive 30546.  This
caused conflicts when we updated to 2.6.35.  I reverted the file to the
official one under the assumption that the original issue has been
addressed by now in the official Linux code.  The original problem had
broken UNH compliance.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>