Piper Wifi: Fixed bug that prevented us from receiving OFDM frames
The semantics of the bss_info_changed callback changed between the 2.6.28
and 2.6.35 kernels.  We used to be able to determine whether the AP was
B only or BG by looking at the basic rate information.  This is no longer
the case and can cause a bug where we disable the OFDM receiver when in
BG mode.  Fixed this by always leaving the OFDM hardware turned on
regardless of what basic rate information is passed to us.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>