Piper Wifi: Fixed issue which caused us to send garbage at start up
The power save function in the wifi driver was attempting to send a null
data frame with the power save bit set to indicate power save is off when
we start the driver.  The problem is that we don't have an AP at that
point and end up sending to address 0:0:0:0:0:0.  Changed the code so that
it will not send to that address.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>