ccwmx51: Correct external clock register bit
When an LCD is configured with internal clock, the ipu register might
be incorrectly configured.

This is because the IPU is configured twice, first for a generic display
with external clock, and then for the configured display. If the IPU
clock is not enabled, the code skips the IPU register configuration for
internal clock. This commit verifies for the internal clock case that
the IPU is correctly configured.

Signed-off-by: Alejandro Gonzalez <alex.gonzalez@digi.com>