ENGR00139455 Add support for pad settings in groups
On MX50 and MX53, there are a group of registers that
control a pad setting for a whole group of pins. This commit
adds:

1-A new define for mapping these pins: IOMUX_PAD_GRP(offset, shift).
The offset is the register offset of the IOMUXC base and the shift
is the bit(s) to be set in the register.

2-A new bitmap, iomux_grp_cfg_t, this is a 32-bit bitmap. The upper
16 bits hold the register offset and the lower 16 bits hold the
shift value.

3-A new function, mxc_iomux_set_pad_groups(iomux_grp_cfg_t pad_grp,
int value). The sample usage is:
  a) in iomux-mx50.h,
   #define MX50_IOMUXC_SW_PAD_CTL_GRP_ADDDS
  b) Call the function like this:
    mxc_iomux_set_pad_groups(MX50_IOMUXC_SW_PAD_CTL_GRP_ADDDS, 0x5);

Signed-off-by: Dinh Nguyen <Dinh.Nguyen@freescale.com