ns9xxx pm/irq: do not disable irqs on suspend
Commit 2ed8d2b3a81bdbb0418301628ccdb008ac9f40b7 mask the interrupts
before going to suspend.
In some (most) architectures, the interrupts can be masked at the PIC so
that when they happen, they are not routed to the CPU. In ns9xxx the
interrupts can only be disabled, so if they are disabled, they cannot
wake the system up.

This commit blocks the disabling of all interrupts during a suspend.
However the right way to do would be:
1) identify WAKEUP interrupts and disable all non-wakeup interrupts.
2) create an 'enable' callback for enabling interrupts disabled at 1.
3) Fix an aparent bug in the disable() callback where the status is cleared
   rather than setting IRQ_SUSPENDED flag.

I tried to do these three operations in 2e80d62ad1890c75b0c4d448ea46be01b8e566c6
but when an RTC interrupt occurred the system was blocked. So I decided to just
wrap the new code that disables the interrupts.

Fixes vantives 37819, 37820, 37821

Signed-off-by: Hector Palacios <hector.palacios@digi.com>