s3c2410_wdt: add support for disabling watchdog
The watchdog can only be disabled if WATCHDOG_NOWAYOUT is not defined.
If it is defined, an error message is printed and an error code returned.

Also remove error message when the watchdog is closed without being
stopped. It is not necessarily an error. For example the
wd_test application closes the watchdog without disabling it.
In this case the error message may be confusing.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
(cherry picked from commit 3d84133310a94f78796916c038486a3ce78dcf09)

Signed-off-by: Hector Palacios <hector.palacios@digi.com>