digi wifi: Changed user interface to power save feature
Changed the user interface to power save so that it is no longer
controlled by mac80211.  This means the commands

   iwconfig wlanx power on
   iwconfig wlanX power off

will no longer work.  Instead, writing a valid duty cycle other
than 100 to the sys file power_duty will turn duty cycling on,
and writing a duty cycle of 100 will turn power save off.

This change was necessary because when mac80211 was updated from
the 2.6.31 release to 2.6.35 it was improved so that it would
disable the power save feature whenever it had data to transfer.
This doesn't work for us because Piper has a bug in it which
causes it to overheat when under load.  Therefore we need to do
duty cycling when under load.  The change takes duty cycling out
of the control of mac80211 and allows the user to control it
directly.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>