nand: added a configuration option to remove the default boot partition
The Freescale platform code for mx28 based platforms creates a 20M boot
partition in NAND.  We want to specify all of the partitions as command
line arguments to the kernel.  I added a configuration option,
CONFIG_REMOVE_DEFAULT_FREESCALE_PARTITION, that makes it possible to
confgure the platform so that no partitions are created by default and
all of the paritioning information is specified on the command line.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>