wifi: updated gpio pin definitions for wifi hardware
I updated the GPIO pin definitions in this file to set up the ATHEROS_PWR and WIFI_ACTIVITY pins.
The driver is suppose to blink the WIFI_ACTIVITY LED when there is activity.  This LED may be
repurposed in the future.  The ATHEROS_PWR is used to power on and off the Atheros wifi chip.

I also removed the pin definitions for sound card since our board does not have a sound card and
the sound driver wanted to use the WIFI_ACTIVITY LED.

Signed-off-by: Charles Gordon <cgordon@digi.com>