cpx2: take already configured MAC from FEC registers
Add a config option to NOT read the MAC address from OTP bits.
Freescale expects the 2 higher bytes of the MAC to be 00:04 and
the 4 lower bytes of the MAC address to be stored in the OTP custom
area. The platform device initialization fills in the ethernet device
platform data with this information.
If the new option is enabled, the platform data MAC is set to zeros and
the driver will check if there is already a valid MAC configured in the
FEC registers. Since this is set by U-Boot it will use that.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>