initramfs: Allow using initramfs with empty /dev
When booting a kernel with no initramfs, there is a basic /dev hardcoded
in noinitramfs.c.

This contains a /dev/console device. Also, DEVTMPFS can be used to have
the kernel mount a tmpfs /dev and have the devices create nodes on it
when they are added.

When using an initramfs there is no embedded /dev in the kernel, and
the DEVTMPFS is mounted after the sys_open call to /dev/console in
kernel_init. Realistically, the kernel expects the initramfs to contain
a /dev/console node.

This patch reuses Digi specific TMPFSDEV which was used before the
DEVTMPFS support existed to allow for a tmpfs to be mounted in /dev with
a console and a null node so that initramfs with an empty /dev folder
can still boot.

Signed-off-by: Alejandro Gonzalez <alex.gonzalez@digi.com>