nx9xxx_wdt: fixed ns921x clock source for watchdog
In ns921x processors, the watchdog is not sourced from the CPU
clock but from AHB clock.
The final frequency is determined by 2 * AHB clock frequency.

NOTE: The watchdog clock frequency of (2 * AHB) is not correctly
described in the hardware reference manual (90000846_D/90000847_D).
This was taken from NET+OS code and apparently was "inside information"
that was not updated in the HRM.

This solves vantive case #38531

Signed-off-by: Hector Palacios <hector.palacios@digi.com>