mx28: Allow for pins to be configured without a previous request.
A pin that is going to be used for GPIO and is going to be controlled
through the sysfs needs to allow for PAD configuration without an explicit
kernel request. If the kernel requests it then user space will not be
able to.

This wouldn't be needed once the PINMUX support is included in the kernel
on the 3.0 kernel series, as this would allow for PAD configuraton from
user space.

This commit also configures the SSP0_DETECT pin used in the X2 pushbutton
to have the internal pull up enabled.

Signed-off-by: Alejandro Gonzalez <alex.gonzalez@digi.com>