ns9xxx-eth: put the Ethernet PHY to power down on suspend
When suspending, the Ethernet clock was disabled but the PHY remained
powered and consuming.
This commit writes a command to the PHY to put it in POWER DOWN mode.

In the cme9210, this saves 95mA @ 3.3VDC.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
(cherry picked from commit 550c9e9e4fa0e2b2556c9d7770b49d3846915b1d)