wifi:  disable SPP_DETECT pin
Found the enabling the SPP_DETECT pin causes a conflict on the cpx2 board.  We can no longer control
the network LED, and the console serial port acts strangely.  The SPP_DETECT pin is normally used by
the MMC/SDIO driver, which is needed to communicate with the Atheros chip.  However, it appears that
the driver can work without this pin, so I changed the code to leave it disabled.

Signed-off-by: Charles Gordon <cgordon@digi.com>