2443-serial: rs485 flow control support
This patch controls RTS line as GPIO to generate RS485 half duplex
flow control. RTS is activated (configurable level, default active high)
when the TX unit and interrupt are enabled to transmit data.

The TX ISR schedules a tasklet that will deactivate RTS line when
TX unit and interrupt are disabled and buffer is empty and all data have
been transmitted out of the shift register.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>