ns9xxx-i2c: fix calculation of i2c clock
The formulas to calculate CLREF bitfield of i2c configuration register
were wrong and have been fixed. The following problems existed:
  1. The VSCD field was set to 1 causing a divisor to be applied which
     resulted in a frequency around 16KHz (apparently a divisor by 8).
  2. The TMDE bitfield was not correctly programmed depending on the
     frequency.
  3. The scl_delay value was hardcoded for all platforms and resulted
     in inaccurate frequencies.

NetSilicon doesn't know either about the correct formulas so empirical
measurements and calculations were done to get to scl_delay values that can
generate the most accurate signals for each platform.

cme9210js platform was not able to generate an appropriate 400KHz signal
so this frequency is unallowed for this platform.

This fixes vantive #39091.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>