fim sdio: fixed Vantive #38541, firmware reload code didn't work
Problem:
The FIM-SDIO driver was compiled as a module. After the modprobe, the
firmware stopped working, and after a card was plugged in, it crashed the
whole system.

Solution:
The main problem was that an old FIM fimware was in the rootfs, and it
couldn't communicate properly with the driver. But the firmware reload
crashed the system, when it shouldn't do that. It was because the driver
didn't pass the SDIO FIM firmware's name to the FIM layer.

Signed-off-by: Robert Hodaszi <robert.hodaszi@digi.com>