cpx2 wifi:  corrected gpio pin configuration for cpx2 wifi
The original GPIO pin configuration for the wifi version of the product
was incorrect.  The wifi build uses the SDIO/MMC interface to
communicate with the Atheros wifi chip.  The original configuration
enabled all of the SDIO/MMC pins, however, only a few of them are used.
Data pins 4-7 are not connected and should be left disabled.  The
SSP0_DETECT pin is also not used for SDIO, and is actually used to
control one of the LEDs on the board and will interfere with the console
serial port unless it is configured as in input.

I corrected the GPIO configuration settings.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>