mxc_nd2: calculate limit page of boot part to skip BI swap
For platforms that need to skip the BI swap on the boot partition,
calculate the limit page to skip the swap basing on the first
partition size and the flash page size.

The partition size is retrieved from the partition table passed in the
command line. This has to be calculated before the second phase of the
flash detection, where the bad blocks are scanned. For this reason the
nand_scan_tail() is moved after calculating the boot part size, so that
we can skip the BI swap (if needed) before scanning for bad blocks.

This patch removes the previous hardcoded values, which only apply if
the partition table is not given.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>