MX5X: IPU: Check display boundaries
Sometimes the display number is set as -1 to mark the overlay layer. Not
checking the validity of the display input was causing a kernel oops and
a user space freeze when stopping the playback of hardware accelerated
media files.

This patch checks whether the display number passed is negative and if so
does nothing in those functions trying to hash into a display table. You
could argue an upper bound check could also be useful, but there seems
to be no cases where the display is set to such value so the check would
be a bit paranoid.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>