2443 gpio: constrain return value of gpio_get to 0|1
The s3c2443_gpio_get() function was returning the value of the
gpio read register instead of a 0 or a 1, for a gpio.
Applications like the gpio_test that check for a 0 or a 1, may fail
if they are receiving a positive value bigger than 1.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
(cherry picked from commit 1ea91297c2e3c03830f425eaeca04f4d1e2f499a)