2443 audio: for 2443 iis, report we only support 44.1KHz
Report we only support the frequency of 44100 Hz so that the driver
adjusts the codec to play the different frequencies. If we report we
support other frequencies, some may play correctly but some others
(8000, 32000) play incorrectly, for example:
 - 8000 plays with bad quality (as if less samples were played)
 - 32000 plays slowly (10% slower than it should)

The most correct would be to report the frequencies that is able to
calculate correctly. With only one frequency, the LRCK (word select
or channel select) signal is fixed to that frequency and the driver
does internal adjustments to play at the correct frequency.

This fixes vantive 39786.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>