fim-sdio: fix firmware restart / removing the SD-card without unmount could cause invalid driver state
Firmware restart function was not worked before -> fixed.

If the user removed the SD-card under communication or without unmount, the
driver got into an invalid state. It was because e.g. in case of a transmit
error, the FIM firmware interrupted the communication after the problematic
block, but the driver sent more blocks in a DMA buffer, so the FIM interpreted
the remaining data as new commands.
It is fixed in the FIM firmware also, see in the FIM repository:
bf7188b061315ce9c8cd56c8cc37414851fe3556

In the fim_sd_send_command() function, the command response validater routine
moved before the block writing. It causes a shorter command timeout time.

Signed-off-by: Robert Hodaszi <robert.hodaszi@digi.com>