mxc_camera: fix support for dual camera
Support for the two cameras was incomplete and caused the
camera2 be not detected if the camera1 was not connected
(see vantive 40526). Also, two devices (video0 and video1)
were created no matter if only one camera was enabled in the
kernel config.

The commit, wraps the different arrays for the two camera sensors
into the kernel config defines for each camera. The function that
returns the CSI of the camera is substituted to return the index
of the camera in the arrays, instead.
The CSI of the camera is retrieved from field 'id' of the
mxc_v4l2_devices array.

The cameras (if enabled in kernel config) are registered, no matter
if the sensor is detected, otherwise camera2 would not be detected
if camera1 is not connected.

This fixes vantive 40526.

NOTE: a race appeared between function init_camera_struct() in
mxc_v4l2_capture.c and function mt9v111_probe() in mt9v111.c.
The first function registers the device which calls the probe.
Apparently the race happened when accessing the i2c_client pointer.
Moving this instruction below in the probe() made the race disappear
but it might show up again in the future.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
(cherry picked from commit 870484c800401ab91945af30713756fd292a313f)