cpx2: Add support for digi_product_id=
This adds support for a new command-line argument
'digi_product_id=' which is setup in manufacturing to
be passed in via ubootenv. This was available already in
user-mode using 'prodinfoenv', but I'm adding it for
kernel-mode in case we ever need to distinguish between
cpx2 variants in the platform startup code. Large changes
should of course be defined as new platforms, but this
provides a fail-safe should we ever need a slight change
in the kernel code based on the product_id.

Signed-off-by: Steve Klein <stevek@digi.com>