CCXMX51: Add VPU IRAM configuration option
This option was moved to platform code in ENGR00140893-1 (9672393), in
order to have a different value between MX51 and MX53 on freescale's
multiplatform kernel.

Freescale's recommendation is to disable it in the MX51 and enable it on
the MX53.

However, on our testing we have found it working better on some conditions
with the IRAM enabled on the MX51.

This commit reintroduces the option as a configuration element, and defaults
it to enabled as that was the DEL 5.6 default.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>