mx5-usb: return a valid wakeup event code when OWIR not signaled
If the USB OTG cable is connected together with a USB disk while
the target is suspended, an interrupt takes place when the system
is resumed, but the OWIR (OTG wake up interrupt request) flag is
not set, so an invalid code is returned and the system enters
and endless loop where the interrupt is set but never cleared.

For this special case, we will return a valid ID change code. The
result is that the USB disk interrupt is not recognized (and the disk
not mounted) but the target is able to wake from sleep.
Disconnecting and connecting the USB disk again makes the system
detect the disk again.

This works around 40541.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
(cherry picked from commit dc7c2eefca7d2c819902d634d7e105f6922a5833)