mma7455: Modify I2C read and write implementation.
The mma7455 driver is not working with the IMX_I2C driver, but it is
with the MXC_I2C one.

This commit modifies the I2C send and receive driver functions to
use i2c_master_send() and i2c_master_rcve() instead of directly
i2c_transfer. This change makes the driver work with both I2C bus drivers.

This fixes Vantive 40023.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>