cc9m2443: IIS, invert LRC polarity
This is a similar fix to the one for the CCWMX51 with the peculiarity that
the IIS interface sets the polarity of the LRC when being configured
for the serial data format. It does not offer an API to modify the
LRC polarity. The fix inverts the LRC polarity for the CC9M2443 when
configuring the I2S mode.

This fixes the channel inversion both in playback and recording and
fixes v39741.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>