mma7455l: do not output DRDY pin on Measurement mode
When setting the mode to Measurement (0x01 to Mode Control Register
at address 0x16), the DRDY pin was by default output to INT1/DRDY pin
causing an interrupt. It looks like the interrupt is not correctly
handled and hangs the system.

To avoid this, when setting any mode set also DRPD flag to
not output Data Ready status to DRDY/INT1 pin.

The interrupt logic must be reviewed because it hangs the system.
Currently, this driver does not correctly support interrupts.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>