ccxmx5x ipu_disp: Fix problem of 1 pixel shifted and fuzzy VGA
This commit implements a workarround to avoid that the video output
is shifted 1 pixel to the right.
This happens when the original code from FSL (after applying a patch)
and has been also observed in their Evaluation Board.
Their fix tries to correct a fuzzy VGA output.
In order to get the best video output we only apply the new code when
the configured output is the VGA interface.
This is the best fix I can implement at this point. This is the video
controller microcode and I dont know what they are exactly doing in
each configuration. We have to check with FSL for a better fix.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>